+++ app/src/main/java/com/marcomadalin/olympus/data/database/entities/ExerciseEntity.kt
@@ -8,9 +8,9 @@ import com.marcomadalin.olympus.domain.model.Exercise
 @Entity(tableName = "Exercises")
 data class ExerciseEntity(
     @PrimaryKey(autoGenerate = true)
-    var id : Int = 0,
-    var workoutId : Int = 0,
-    var exerciseDataId : Int = 0,
+    var id : Long = 0,
+    var workoutId : Long = 0,
+    var exerciseDataId : Long = 0,
     var restTime : Long = 0,
     var note : String = "",
     var exerciseNumber : Int = 0
