+++ app/src/main/java/com/marcomadalin/olympus/data/database/entities/RoutineEntity.kt
@@ -7,8 +7,8 @@ import com.marcomadalin.olympus.domain.model.Routine
 @Entity(tableName = "Routines")
 open class RoutineEntity (
     @PrimaryKey(autoGenerate = true)
-    open var id : Int = 0,
-    open var userId : Int = 0,
+    open var id : Long = 0,
+    open var userId : Long = 0,
     open var name : String = "",
     open var note : String = "",
 )
