+++ app/src/main/java/com/marcomadalin/olympus/domain/model/Exercise.kt
@@ -5,9 +5,9 @@ import java.time.Duration
 
 
 data class Exercise(
-    var id : Int = 0,
-    var workoutId : Int = 0,
-    var exerciseDataId : Int = 0,
+    var id : Long = 0,
+    var workoutId : Long = 0,
+    var exerciseDataId : Long = 0,
     var restTime : Duration = Duration.ofSeconds(0),
     var note : String = "",
     var exerciseNumber : Int = 0,
