+++ app/src/main/java/com/marcomadalin/olympus/domain/model/Measure.kt
@@ -5,8 +5,8 @@ import com.marcomadalin.olympus.domain.model.enums.MeasurePart
 import java.time.LocalDate
 
 data class Measure(
-    var id: Int = 0,
-    var userId: Int = 0,
+    var id: Long = 0,
+    var userId: Long = 0,
     var date: LocalDate = LocalDate.parse(""),
     var value: Double = 0.0,
     var part: MeasurePart = MeasurePart.Weight,
